Add meson build system project
Defined the project so that the meson build system can build the console
emulator. Numerous comments explain what each function does and how to
use the project file to build it afterwards.
(Also adds a .gitignore file for the .vscode folder, on account of not
needing to have that on vcs.)